feat(card): add border radius tokens
DSY-2549
